feat: add AdminViewOfCustomerAccountsRepository.cs

+++ src/Lab5/Lab5.Application.Abstractions/Repositories/ICustomerAccountsRepository.cs
@@ -4,6 +4,6 @@ namespace Application.Abstractions.Repositories;
 
 public interface ICustomerAccountsRepository
 {
-    string? FindAccountPasswordById(long userId, long accountId);
-    CustomerAccount? FindAccountById(long userId, long accountId);
+    string? FindAccountPasswordById(long userId, decimal accountId);
+    CustomerAccount? FindAccountById(long userId, decimal accountId);
 }
